refactor(BasicAcceptance): Acceptance 에서 초기 셋팅하는 부분을 class 로 추출
